Moved color class to mat.py and fixed modules to work with new opencv wrapper
