+++ opsi/modules/nt.py
@@ -1,10 +1,12 @@
 from dataclasses import dataclass
 
 from networktables import NetworkTables
+
 from opsi.manager.manager_schema import Function, Hook
 from opsi.manager.netdict import NetworkDict
-from opsi.manager.types import AnyType, Color
+from opsi.manager.types import AnyType
 from opsi.util.cv import Point
+from opsi.util.cv.mat import Color
 from opsi.util.networking import get_nt_server
 from opsi.util.unduplicator import Unduplicator
